app/requirements/base.txt
asgiref==3.9.1
dj-database-url==3.0.1
Django==5.2.4
python-dotenv==1.1.1
sqlparse==0.5.3
tzdata==2025.2
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.0
pyjwt==2.9.0